Update Nth_Fibonacci_Number.java
This is a solution which uses 2*2 matrices. 
1 1
1 0 
We start with the above matrix. In this matrix, element at (0,0) position gives first (n+1)th fibonacci number and element at (1,1) position gives nth fibonacci number.
We multiply this matrix with itself or with identity matrix depending upon the value of N each time.